chatbot/requirements.txt
# To Install run pip install -r requirements.txt


# FastAPI for building APIs
fastapi

# Pydantic for data validation
pydantic

# Typing support (included in Python 3.8+, typing_extensions for backward compatibility)
typing_extensions

# LangChain and OpenAI / DeepSeek integrations
langchain
langchain-openai
langchain-deepseek

# Token encoding (used in LangChain/OpenAI)
tiktoken

# PostgreSQL support
psycopg2-binary

# SQLAlchemy ORM
sqlalchemy

# NumPy for embedding/vector ops
numpy

# HTTP requests
requests

# .env management
python-dotenv

langchain-community

pandas

# Optional: FAISS or Chroma for vector database (only if used)
# faiss-cpu
# chromadb
